topic_analysis/backend/legacy_backup/requirements-fast.txt
# Fast requirements for TrendTap backend (without heavy dependencies)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
structlog==23.2.0
psutil==5.9.6
prometheus-client==0.19.0
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0


